cmd/go, cmd/compile: add communication pipe to signal export data is ready
Now the compiler has a new -exportfd flag, which if set is a fd the
compiler should write a byte to once export data is ready. That
allows the go command to unblock all the dependent compiles. We added
a shellCmd struct to hold the state of the compile in progress so that
the build (backend) action can wait on it for the compile to end.

For #15734

Reviewed-on: https://go-review.googlesource.com/c/go/+/828184
LUCI-TryBot-Result: golang-scoped@luci-project-accounts.iam.gserviceaccount.com <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>

+++ src/cmd/compile/internal/base/print.go
@@ -263,7 +263,7 @@ func AssertfAt(b bool, pos src.XPos, format string, args ...any) {
 func hcrash() {
 	if Flag.LowerH != 0 {
 		FlushErrors()
-		if Flag.LowerO != "" {
+		if Flag.LowerO != "" && Flag.ExportFD <= 0 {
 			os.Remove(Flag.LowerO)
 		}
 		panic("-h")
@@ -274,7 +274,7 @@ func hcrash() {
 // It flushes any pending errors, removes the output file, and exits.
 func ErrorExit() {
 	FlushErrors()
-	if Flag.LowerO != "" {
+	if Flag.LowerO != "" && Flag.ExportFD <= 0 {
 		os.Remove(Flag.LowerO)
 	}
 	os.Exit(2)
